backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
transformers==4.46.0
torch>=2.5.0
torchvision>=0.20.0
pillow==11.0.0
pdf2image==1.17.0
opencv-python==4.10.0.84
numpy==2.1.0
rapidfuzz==3.10.0
python-Levenshtein==0.26.0
pydantic==2.10.0
qrcode[pil]==8.0
cryptography==44.0.0
python-dateutil==2.9.0
pytesseract==0.3.13
